# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/store-visits/{id}": {
      "get": {
        "summary": "Get a store visit by ID",
        "tags": [
          "Store Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store visit details"
          }
        }
      },
      "put": {
        "summary": "Update a store visit",
        "tags": [
          "Store Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store visit updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a store visit",
        "tags": [
          "Store Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store visit deleted"
          }
        }
      }
    },
    "/api/store-visits": {
      "get": {
        "summary": "Get all store visits",
        "tags": [
          "Store Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of store visits"
          }
        }
      },
      "post": {
        "summary": "Create a store visit",
        "tags": [
          "Store Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "owner_contact",
                  "location",
                  "date"
                ],
                "properties": {
                  "owner_contact": {
                    "type": "string"
                  },
                  "number_of_visits": {
                    "type": "integer"
                  },
                  "location": {
                    "type": "string"
                  },
                  "time": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Store visit created"
          }
        }
      }
    },
    "/api/website-visits/{id}": {
      "get": {
        "summary": "Get a website visit by ID",
        "tags": [
          "Website Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Website visit details"
          },
          "404": {
            "description": "Website visit not found"
          }
        }
      },
      "put": {
        "summary": "Update a website visit",
        "tags": [
          "Website Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Website visit updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a website visit",
        "tags": [
          "Website Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Website visit deleted"
          }
        }
      }
    },
    "/api/website-visits": {
      "get": {
        "summary": "Get all website visits",
        "tags": [
          "Website Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of website visits"
          }
        }
      },
      "post": {
        "summary": "Create a website visit",
        "tags": [
          "Website Visits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ip",
                  "date"
                ],
                "properties": {
                  "ip": {
                    "type": "string"
                  },
                  "owner_contact": {
                    "type": "string"
                  },
                  "number_of_visits": {
                    "type": "integer"
                  },
                  "page_visits": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "website_duration": {
                    "type": "integer"
                  },
                  "location": {
                    "type": "string"
                  },
                  "time": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Website visit created"
          }
        }
      }
    }
  },
...
}